+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Possible log types:
 ### Unreleased
 
 - [changed] Bump rust toolchain to 1.86
+- [changed] Change type of `project_id` to String
 
 ### [5.0.4][v5.0.4] (2025-05-16)